Added year of release back in
Not sure how, but year of release had gone missing from the movies
ontology. Have added it back in using dc:date. We may need to alter
this in future as dc:date is designed to refer to any date, rather than
just years. Alternatively, we could capture the exact release date for
movie in the UK market instead.